requirements.txt
fastapi
uvicorn
rembg
python-multipart
pillow
numpy